requirements.txt
glean-indexing-sdk== 0.2.0
glean-api-client==0.7.1
python-dotenv
requests